Changed namespace of Scenarios test and added test for rules matching
