+++ tests/Rules.Framework.IntegrationTests.Common/Features/Stubs/ConditionType.cs
@@ -1,0 +1,11 @@
+namespace Rules.Framework.Tests.Stubs
+{
+    public enum ConditionType
+    {
+        ConditionType1 = 1,
+
+        ConditionType2 = 2,
+
+        ConditionType3 = 3
+    }
+}
