+++ tests/Rules.Framework.IntegrationTests.Common/Features/Stubs/ConditionTypeClass.cs
@@ -1,0 +1,9 @@
+namespace Rules.Framework.Tests.Stubs
+{
+    internal class ConditionTypeClass
+    {
+        public int DummyConditionTypeProperty1 { get; set; }
+
+        public string DummyConditionTypeProperty2 { get; set; }
+    }
+}
