+++ tests/Rules.Framework.IntegrationTests.Common/Features/Stubs/ContentType.cs
@@ -1,0 +1,9 @@
+namespace Rules.Framework.Tests.Stubs
+{
+    public enum ContentType
+    {
+        ContentType1 = 1,
+
+        ContentType2 = 2
+    }
+}
