+++ tests/Rules.Framework.IntegrationTests.Common/Features/Stubs/ContentTypeClass.cs
@@ -1,0 +1,9 @@
+namespace Rules.Framework.Tests.Stubs
+{
+    public class ContentTypeClass
+    {
+        public int DummyContentTypeProperty1 { get; set; }
+
+        public string DummyContentTypeProperty2 { get; set; }
+    }
+}
