+++ tests/Rules.Framework.IntegrationTests.Common/Features/Stubs/EmptyContentType.cs
@@ -1,6 +1,6 @@
 namespace Rules.Framework.Tests.Stubs
 {
-    internal enum EmptyContentType
+    public enum EmptyContentType
     {
     }
 }
