+++ tests/Rules.Framework.IntegrationTests/Scenarios/Scenario2/CarInsuranceAdvisorTests.cs
@@ -1,69 +1,89 @@
-namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario2
+namespace Rules.Framework.IntegrationTests.Scenarios.Scenario2
 {
     using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
     using FluentAssertions;
-    using Microsoft.Extensions.DependencyInjection;
-    using Rules.Framework;
     using Rules.Framework.Builder;
     using Rules.Framework.Core;
     using Rules.Framework.IntegrationTests.Common.Scenarios.Scenario2;
-    using Rules.Framework.Providers.InMemory;
     using Xunit;
 
-    public class CarInsuranceAdvisorTests : BaseScenarioTests
+    public class CarInsuranceAdvisorTests
     {
-        private readonly InMemoryRulesStorage<ContentTypes, ConditionTypes> inMemoryRulesStorage;
-
-        public CarInsuranceAdvisorTests()
-        {
-            this.inMemoryRulesStorage = new InMemoryRulesStorage<ContentTypes, ConditionTypes>();
-            this.LoadInMemoryStorage<ContentTypes, ConditionTypes, CarInsuranceAdvices>(
-                DataSourceFilePath,
-                this.inMemoryRulesStorage,
-                (c) => this.Parse<CarInsuranceAdvices>((string)c));
-        }
-
         private static string DataSourceFilePath => $@"{Environment.CurrentDirectory}/Scenarios/Scenario2/rules-framework-tests.car-insurance-advisor.json";
 
         [Fact]
-        public async Task GetCarInsuranceAdvice_RepairCostsNotWorthIt_ReturnsPayOldCar()
+        public async Task GetCarInsuranceAdvice_ClaimDescriptionContionsAlcoholOrDrugs_ReturnsPerformInvestigation()
         {
             // Arrange
-            CarInsuranceAdvices expected = CarInsuranceAdvices.PayOldCar;
+            CarInsuranceAdvices expected = CarInsuranceAdvices.PerformInvestigation;
             const ContentTypes expectedContent = ContentTypes.CarInsuranceAdvice;
-            DateTime expectedMatchDate = new DateTime(2016, 06, 01, 20, 23, 23);
+            DateTime expectedMatchDate = new DateTime(2020, 06, 01);
             Condition<ConditionTypes>[] expectedConditions = new Condition<ConditionTypes>[]
             {
                 new Condition<ConditionTypes>
                 {
                     Type = ConditionTypes.RepairCosts,
-                    Value = 0.0m
+                    Value = 800.00000m
                 },
                 new Condition<ConditionTypes>
                 {
                     Type = ConditionTypes.RepairCostsCommercialValueRate,
-                    Value = 0.0m
+                    Value = 23.45602m
+                },
+                new Condition<ConditionTypes>
+                {
+                    Type = ConditionTypes.ClaimDescription,
+                    Value = "Driver A claims that Driver B appeared to be under the effect of alcohol."
                 }
             };
 
-            IServiceCollection serviceDescriptors = new ServiceCollection();
-            serviceDescriptors.AddSingleton(this.inMemoryRulesStorage);
-            IServiceProvider serviceProvider = serviceDescriptors.BuildServiceProvider();
+            IRulesDataSource<ContentTypes, ConditionTypes> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<ContentTypes, ConditionTypes>(DataSourceFilePath, serializedContent: false);
 
             RulesEngine<ContentTypes, ConditionTypes> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<ContentTypes>()
                 .WithConditionType<ConditionTypes>()
-                .SetInMemoryDataSource(serviceProvider)
+                .SetDataSource(rulesDataSource)
                 .Configure(reo =>
                 {
                     reo.PriotityCriteria = PriorityCriterias.BottommostRuleWins;
                 })
                 .Build();
 
+            RuleBuilderResult<ContentTypes, ConditionTypes> ruleBuilderResult = RuleBuilder.NewRule<ContentTypes, ConditionTypes>()
+                .WithName("Car Insurance Advise on on accident under the effect of drugs or alcohol")
+                .WithDateBegin(DateTime.Parse("2020-01-01"))
+                .WithCondition(b =>
+                {
+                    return b.AsComposed()
+                        .WithLogicalOperator(LogicalOperators.Or)
+                        .AddCondition(cb =>
+                        {
+                            return cb.AsValued(ConditionTypes.ClaimDescription)
+                                .OfDataType<string>()
+                                .WithComparisonOperator(Operators.Contains)
+                                .SetOperand("alcohol")
+                                .Build();
+                        })
+                        .AddCondition(cb =>
+                        {
+                            return cb.AsValued(ConditionTypes.ClaimDescription)
+                                .OfDataType<string>()
+                                .WithComparisonOperator(Operators.Contains)
+                                .SetOperand("drugs")
+                                .Build();
+                        })
+                        .Build();
+                })
+                .WithContentContainer(new ContentContainer<ContentTypes>(expectedContent, t => CarInsuranceAdvices.PerformInvestigation))
+                .Build();
+
             // Act
+            await rulesEngine.AddRuleAsync(ruleBuilderResult.Rule, RuleAddPriorityOption.AtBottom).ConfigureAwait(false);
+
             Rule<ContentTypes, ConditionTypes> actual = await rulesEngine.MatchOneAsync(expectedContent, expectedMatchDate, expectedConditions);
 
             // Assert
@@ -93,14 +113,13 @@ namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario2
                 }
             };
 
-            IServiceCollection serviceDescriptors = new ServiceCollection();
-            serviceDescriptors.AddSingleton(this.inMemoryRulesStorage);
-            IServiceProvider serviceProvider = serviceDescriptors.BuildServiceProvider();
+            IRulesDataSource<ContentTypes, ConditionTypes> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<ContentTypes, ConditionTypes>(DataSourceFilePath, serializedContent: false);
 
             RulesEngine<ContentTypes, ConditionTypes> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<ContentTypes>()
                 .WithConditionType<ConditionTypes>()
-                .SetInMemoryDataSource(serviceProvider)
+                .SetDataSource(rulesDataSource)
                 .Configure(reo =>
                 {
                     reo.PriotityCriteria = PriorityCriterias.BottommostRuleWins;
@@ -114,6 +133,93 @@ namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario2
             actual.Should().NotBeNull();
             CarInsuranceAdvices actualContent = actual.ContentContainer.GetContentAs<CarInsuranceAdvices>();
             actualContent.Should().Be(expected);
+        }
+
+        [Fact]
+        public async Task GetCarInsuranceAdvice_SearchForRulesExcludingRulesWithoutSearchConditions_ReturnsNoRules()
+        {
+            // Arrange
+            const ContentTypes expectedContent = ContentTypes.CarInsuranceAdvice;
+            DateTime expectedMatchDate = new DateTime(2018, 06, 01);
+            SearchArgs<ContentTypes, ConditionTypes> searchArgs = new SearchArgs<ContentTypes, ConditionTypes>(expectedContent, expectedMatchDate, expectedMatchDate)
+            {
+                Conditions = new Condition<ConditionTypes>[]
+                {
+                    new Condition<ConditionTypes>
+                    {
+                        Type = ConditionTypes.RepairCosts,
+                        Value = 800.00000m
+                    },
+                    new Condition<ConditionTypes>
+                    {
+                        Type = ConditionTypes.RepairCostsCommercialValueRate,
+                        Value = 86.33m
+                    }
+                },
+                ExcludeRulesWithoutSearchConditions = true
+            };
+
+            IRulesDataSource<ContentTypes, ConditionTypes> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<ContentTypes, ConditionTypes>(DataSourceFilePath, serializedContent: false);
+
+            RulesEngine<ContentTypes, ConditionTypes> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
+                .WithContentType<ContentTypes>()
+                .WithConditionType<ConditionTypes>()
+                .SetDataSource(rulesDataSource)
+                .Configure(reo =>
+                {
+                    reo.PriotityCriteria = PriorityCriterias.BottommostRuleWins;
+                })
+                .Build();
+
+            // Act
+            IEnumerable<Rule<ContentTypes, ConditionTypes>> actual = await rulesEngine.SearchAsync(searchArgs);
+
+            // Assert
+            actual.Should().NotBeNull();
+            actual.Should().HaveCount(0);
+        }
+
+        [Fact]
+        public async Task GetCarInsuranceAdvice_SearchForRulesWithRepairCostsGreaterThan1000_Returns2Rules()
+        {
+            // Arrange
+            const ContentTypes expectedContent = ContentTypes.CarInsuranceAdvice;
+            DateTime expectedMatchDate = new DateTime(2018, 06, 01);
+            SearchArgs<ContentTypes, ConditionTypes> searchArgs = new SearchArgs<ContentTypes, ConditionTypes>(expectedContent, expectedMatchDate, expectedMatchDate)
+            {
+                Conditions = new Condition<ConditionTypes>[]
+                {
+                    new Condition<ConditionTypes>
+                    {
+                        Type = ConditionTypes.RepairCosts,
+                        Value = 1200.00000m
+                    }
+                },
+                ExcludeRulesWithoutSearchConditions = false
+            };
+
+            IRulesDataSource<ContentTypes, ConditionTypes> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<ContentTypes, ConditionTypes>(DataSourceFilePath, serializedContent: false);
+
+            RulesEngine<ContentTypes, ConditionTypes> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
+                .WithContentType<ContentTypes>()
+                .WithConditionType<ConditionTypes>()
+                .SetDataSource(rulesDataSource)
+                .Configure(reo =>
+                {
+                    reo.PriotityCriteria = PriorityCriterias.BottommostRuleWins;
+                })
+                .Build();
+
+            // Act
+            IEnumerable<Rule<ContentTypes, ConditionTypes>> actual = await rulesEngine.SearchAsync(searchArgs);
+
+            // Assert
+            actual.Should().NotBeNull();
+            actual.Should().HaveCount(2);
+            actual.Should().Contain(r => r.Name == "Car Insurance Advise on repair costs greater than 80% of commercial value");
+            actual.Should().Contain(r => r.Name == "Car Insurance Advise on repair costs lesser than 80% of commercial value");
         }
 
         [Fact]
@@ -136,26 +242,22 @@ namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario2
                 }
             };
 
-            IServiceCollection serviceDescriptors = new ServiceCollection();
-            serviceDescriptors.AddSingleton(this.inMemoryRulesStorage);
-            IServiceProvider serviceProvider = serviceDescriptors.BuildServiceProvider();
+            IRulesDataSource<ContentTypes, ConditionTypes> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<ContentTypes, ConditionTypes>(DataSourceFilePath, serializedContent: false);
 
             RulesEngine<ContentTypes, ConditionTypes> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<ContentTypes>()
                 .WithConditionType<ConditionTypes>()
-                .SetInMemoryDataSource(serviceProvider)
+                .SetDataSource(rulesDataSource)
                 .Configure(reo =>
                 {
                     reo.PriotityCriteria = PriorityCriterias.BottommostRuleWins;
                 })
                 .Build();
 
-            IRulesDataSource<ContentTypes, ConditionTypes> rulesDataSource = CreateRulesDataSourceTest<ContentTypes, ConditionTypes>(this.inMemoryRulesStorage);
-
             RuleBuilderResult<ContentTypes, ConditionTypes> ruleBuilderResult = RuleBuilder.NewRule<ContentTypes, ConditionTypes>()
                 .WithName("Car Insurance Advise on self damage coverage")
                 .WithDateBegin(DateTime.Parse("2018-01-01"))
-                .WithPriority(1)
                 .WithContentContainer(new ContentContainer<ContentTypes>(ContentTypes.CarInsuranceAdvice, (t) => CarInsuranceAdvices.Pay))
                 .Build();
             IEnumerable<Rule<ContentTypes, ConditionTypes>> existentRules1 = await rulesDataSource.GetRulesByAsync(new RulesFilterArgs<ContentTypes>
