+++ tests/Rules.Framework.IntegrationTests/Scenarios/Scenario3/BuildingSecuritySystemControlTests.cs
@@ -1,62 +1,17 @@
-namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
+namespace Rules.Framework.IntegrationTests.Scenarios.Scenario3
 {
     using System;
     using System.Collections.Generic;
-    using System.Dynamic;
-    using System.IO;
     using System.Linq;
-    using System.Reflection;
     using System.Threading.Tasks;
     using FluentAssertions;
-    using MongoDB.Driver;
-    using Newtonsoft.Json;
     using Rules.Framework.Core;
     using Rules.Framework.IntegrationTests.Common.Scenarios.Scenario3;
-    using Rules.Framework.Providers.MongoDb;
-    using Rules.Framework.Providers.MongoDb.DataModel;
     using Xunit;
 
-    public sealed class BuildingSecuritySystemControlTests : IDisposable
+    public class BuildingSecuritySystemControlTests
     {
-        private readonly IMongoClient mongoClient;
-        private readonly MongoDbProviderSettings mongoDbProviderSettings;
-
-        public BuildingSecuritySystemControlTests()
-        {
-            this.mongoClient = CreateMongoClient();
-            this.mongoDbProviderSettings = CreateProviderSettings();
-
-            Stream? rulesFile = Assembly.GetExecutingAssembly()
-                .GetManifestResourceStream("Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3.rules-framework-tests.security-system-actionables.json");
-
-            IEnumerable<RuleDataModel> rules;
-            using (StreamReader streamReader = new StreamReader(rulesFile ?? throw new InvalidOperationException("Could not load rules file.")))
-            {
-                string json = streamReader.ReadToEnd();
-
-                IEnumerable<RuleDataModel> array = JsonConvert.DeserializeObject<IEnumerable<RuleDataModel>>(json, new JsonSerializerSettings
-                {
-                    TypeNameHandling = TypeNameHandling.All
-                });
-
-                rules = array.Select(t =>
-                {
-                    SecuritySystemAction securitySystemAction = t.Content.ToObject<SecuritySystemAction>();
-                    dynamic dynamicContent = new ExpandoObject();
-                    dynamicContent.ActionId = securitySystemAction.ActionId;
-                    dynamicContent.ActionName = securitySystemAction.ActionName;
-                    t.Content = dynamicContent;
-
-                    return t;
-                }).ToList();
-            }
-
-            IMongoDatabase mongoDatabase = this.mongoClient.GetDatabase(this.mongoDbProviderSettings.DatabaseName);
-            mongoDatabase.DropCollection(this.mongoDbProviderSettings.RulesCollectionName);
-            IMongoCollection<RuleDataModel> mongoCollection = mongoDatabase.GetCollection<RuleDataModel>(this.mongoDbProviderSettings.RulesCollectionName);
-
-            mongoCollection.InsertMany(rules);
-        }
+        private static string DataSourceFilePath => $@"{Environment.CurrentDirectory}/Scenarios/Scenario3/rules-framework-tests.security-system-actionables.json";
 
         [Fact]
         public async Task BuildingSecuritySystem_FireScenario_ReturnsActionsToTrigger()
@@ -84,10 +39,13 @@ namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
                 }
             };
 
+            IRulesDataSource<SecuritySystemActionables, SecuritySystemConditions> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<SecuritySystemActionables, SecuritySystemConditions>(DataSourceFilePath);
+
             RulesEngine<SecuritySystemActionables, SecuritySystemConditions> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<SecuritySystemActionables>()
                 .WithConditionType<SecuritySystemConditions>()
-                .SetMongoDbDataSource(this.mongoClient, this.mongoDbProviderSettings)
+                .SetDataSource(rulesDataSource)
                 .Build();
 
             // Act
@@ -130,10 +88,13 @@ namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
                 }
             };
 
+            IRulesDataSource<SecuritySystemActionables, SecuritySystemConditions> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<SecuritySystemActionables, SecuritySystemConditions>(DataSourceFilePath);
+
             RulesEngine<SecuritySystemActionables, SecuritySystemConditions> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<SecuritySystemActionables>()
                 .WithConditionType<SecuritySystemConditions>()
-                .SetMongoDbDataSource(this.mongoClient, this.mongoDbProviderSettings)
+                .SetDataSource(rulesDataSource)
                 .Build();
 
             // Act
@@ -175,10 +136,13 @@ namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
                 }
             };
 
+            IRulesDataSource<SecuritySystemActionables, SecuritySystemConditions> rulesDataSource = await RulesFromJsonFile.Load
+                .FromJsonFileAsync<SecuritySystemActionables, SecuritySystemConditions>(DataSourceFilePath);
+
             RulesEngine<SecuritySystemActionables, SecuritySystemConditions> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<SecuritySystemActionables>()
                 .WithConditionType<SecuritySystemConditions>()
-                .SetMongoDbDataSource(this.mongoClient, this.mongoDbProviderSettings)
+                .SetDataSource(rulesDataSource)
                 .Build();
 
             // Act
@@ -192,19 +156,5 @@ namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
             securitySystemActions.Should().Contain(ssa => ssa.ActionName == "EnableEmergencyLights")
                 .And.HaveCount(1);
         }
-
-        public void Dispose()
-        {
-            IMongoDatabase mongoDatabase = this.mongoClient.GetDatabase(this.mongoDbProviderSettings.DatabaseName);
-            mongoDatabase.DropCollection(this.mongoDbProviderSettings.RulesCollectionName);
-        }
-
-        private static MongoClient CreateMongoClient() => new MongoClient($"mongodb://{SettingsProvider.GetMongoDbHost()}:27017");
-
-        private static MongoDbProviderSettings CreateProviderSettings() => new MongoDbProviderSettings
-        {
-            DatabaseName = "rules-framework-tests",
-            RulesCollectionName = "security-system-actionables"
-        };
     }
 }
