+++ tests/Rules.Framework.Providers.MongoDb.IntegrationTests/Scenarios/Scenario2/rules-framework-tests.car-insurance-advisor.json
