+++ tests/Rules.Framework.Providers.MongoDb.IntegrationTests/Scenarios/Scenario3/BuildingSecuritySystemControlTests.cs
@@ -1,32 +1,62 @@
-namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario3
+namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Scenarios.Scenario3
 {
     using System;
     using System.Collections.Generic;
+    using System.Dynamic;
+    using System.IO;
     using System.Linq;
+    using System.Reflection;
     using System.Threading.Tasks;
     using FluentAssertions;
-    using Microsoft.Extensions.DependencyInjection;
+    using MongoDB.Driver;
     using Newtonsoft.Json;
     using Rules.Framework.Core;
     using Rules.Framework.IntegrationTests.Common.Scenarios.Scenario3;
-    using Rules.Framework.Providers.InMemory;
+    using Rules.Framework.Providers.MongoDb;
+    using Rules.Framework.Providers.MongoDb.DataModel;
     using Xunit;
 
-    public class BuildingSecuritySystemControlTests : BaseScenarioTests
+    public sealed class BuildingSecuritySystemControlTests : IDisposable
     {
-        private readonly InMemoryRulesStorage<SecuritySystemActionables, SecuritySystemConditions> inMemoryRulesStorage;
+        private readonly IMongoClient mongoClient;
+        private readonly MongoDbProviderSettings mongoDbProviderSettings;
 
         public BuildingSecuritySystemControlTests()
         {
-            this.inMemoryRulesStorage = new InMemoryRulesStorage<SecuritySystemActionables, SecuritySystemConditions>();
+            this.mongoClient = CreateMongoClient();
+            this.mongoDbProviderSettings = CreateProviderSettings();
 
-            this.LoadInMemoryStorage<SecuritySystemActionables, SecuritySystemConditions, SecuritySystemAction>(
-                DataSourceFilePath,
-                this.inMemoryRulesStorage,
-                (c) => JsonConvert.DeserializeObject<SecuritySystemAction>((string)c));
+            Stream? rulesFile = Assembly.GetExecutingAssembly()
+                .GetManifestResourceStream("Rules.Framework.Providers.MongoDb.IntegrationTests.Scenarios.Scenario3.rules-framework-tests.security-system-actionables.json");
+
+            IEnumerable<RuleDataModel> rules;
+            using (StreamReader streamReader = new StreamReader(rulesFile ?? throw new InvalidOperationException("Could not load rules file.")))
+            {
+                string json = streamReader.ReadToEnd();
+
+                IEnumerable<RuleDataModel> array = JsonConvert.DeserializeObject<IEnumerable<RuleDataModel>>(json, new JsonSerializerSettings
+                {
+                    TypeNameHandling = TypeNameHandling.All
+                });
+
+                rules = array.Select(t =>
+                {
+                    SecuritySystemAction securitySystemAction = t.Content.ToObject<SecuritySystemAction>();
+                    dynamic dynamicContent = new ExpandoObject();
+                    dynamicContent.ActionId = securitySystemAction.ActionId;
+                    dynamicContent.ActionName = securitySystemAction.ActionName;
+                    t.Content = dynamicContent;
+
+                    return t;
+                }).ToList();
+            }
+
+            IMongoDatabase mongoDatabase = this.mongoClient.GetDatabase(this.mongoDbProviderSettings.DatabaseName);
+            mongoDatabase.DropCollection(this.mongoDbProviderSettings.RulesCollectionName);
+            IMongoCollection<RuleDataModel> mongoCollection = mongoDatabase.GetCollection<RuleDataModel>(this.mongoDbProviderSettings.RulesCollectionName);
+
+            mongoCollection.InsertMany(rules);
         }
-
-        private static string DataSourceFilePath => $@"{Environment.CurrentDirectory}/Scenarios/Scenario3/rules-framework-tests.security-system-actionables.json";
 
         [Fact]
         public async Task BuildingSecuritySystem_FireScenario_ReturnsActionsToTrigger()
@@ -54,14 +84,10 @@ namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario3
                 }
             };
 
-            IServiceCollection serviceDescriptors = new ServiceCollection();
-            serviceDescriptors.AddSingleton(this.inMemoryRulesStorage);
-            IServiceProvider serviceProvider = serviceDescriptors.BuildServiceProvider();
-
             RulesEngine<SecuritySystemActionables, SecuritySystemConditions> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<SecuritySystemActionables>()
                 .WithConditionType<SecuritySystemConditions>()
-                .SetInMemoryDataSource(serviceProvider)
+                .SetMongoDbDataSource(this.mongoClient, this.mongoDbProviderSettings)
                 .Build();
 
             // Act
@@ -104,14 +130,10 @@ namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario3
                 }
             };
 
-            IServiceCollection serviceDescriptors = new ServiceCollection();
-            serviceDescriptors.AddSingleton(this.inMemoryRulesStorage);
-            IServiceProvider serviceProvider = serviceDescriptors.BuildServiceProvider();
-
             RulesEngine<SecuritySystemActionables, SecuritySystemConditions> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<SecuritySystemActionables>()
                 .WithConditionType<SecuritySystemConditions>()
-                .SetInMemoryDataSource(serviceProvider)
+                .SetMongoDbDataSource(this.mongoClient, this.mongoDbProviderSettings)
                 .Build();
 
             // Act
@@ -153,14 +175,10 @@ namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario3
                 }
             };
 
-            IServiceCollection serviceDescriptors = new ServiceCollection();
-            serviceDescriptors.AddSingleton(this.inMemoryRulesStorage);
-            IServiceProvider serviceProvider = serviceDescriptors.BuildServiceProvider();
-
             RulesEngine<SecuritySystemActionables, SecuritySystemConditions> rulesEngine = RulesEngineBuilder.CreateRulesEngine()
                 .WithContentType<SecuritySystemActionables>()
                 .WithConditionType<SecuritySystemConditions>()
-                .SetInMemoryDataSource(serviceProvider)
+                .SetMongoDbDataSource(this.mongoClient, this.mongoDbProviderSettings)
                 .Build();
 
             // Act
@@ -174,5 +192,19 @@ namespace Rules.Framework.Providers.InMemory.IntegrationTests.Tests.Scenario3
             securitySystemActions.Should().Contain(ssa => ssa.ActionName == "EnableEmergencyLights")
                 .And.HaveCount(1);
         }
+
+        public void Dispose()
+        {
+            IMongoDatabase mongoDatabase = this.mongoClient.GetDatabase(this.mongoDbProviderSettings.DatabaseName);
+            mongoDatabase.DropCollection(this.mongoDbProviderSettings.RulesCollectionName);
+        }
+
+        private static MongoClient CreateMongoClient() => new MongoClient($"mongodb://{SettingsProvider.GetMongoDbHost()}:27017");
+
+        private static MongoDbProviderSettings CreateProviderSettings() => new MongoDbProviderSettings
+        {
+            DatabaseName = "rules-framework-tests",
+            RulesCollectionName = "security-system-actionables"
+        };
     }
 }
