+++ tests/Rules.Framework.Tests/Builder/RuleBuilderExtensionsTests.cs
@@ -4,7 +4,7 @@ namespace Rules.Framework.Tests.Builder
     using Moq;
     using Rules.Framework.Builder;
     using Rules.Framework.Core;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class RuleBuilderExtensionsTests
