+++ tests/Rules.Framework.Tests/Builder/SelectorsTests.cs
@@ -4,7 +4,7 @@ namespace Rules.Framework.Tests.Builder
     using FluentAssertions;
     using Moq;
     using Rules.Framework.Builder;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
     using static Rules.Framework.Builder.RulesEngineSelectors;
