+++ tests/Rules.Framework.Tests/ConditionTests.cs
@@ -1,7 +1,7 @@
 namespace Rules.Framework.Tests
 {
     using FluentAssertions;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class ConditionTests
