+++ tests/Rules.Framework.Tests/ConditionTypeExtractorTests.cs
@@ -5,7 +5,7 @@ namespace Rules.Framework.Tests
     using FluentAssertions;
     using Rules.Framework.Core;
     using Rules.Framework.Core.ConditionNodes;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class ConditionTypeExtractorTests
