+++ tests/Rules.Framework.Tests/Core/ConditionNodes/DecimalConditionNodeTests.cs
@@ -3,7 +3,7 @@ namespace Rules.Framework.Tests.Core.ConditionNodes
     using FluentAssertions;
     using Rules.Framework.Core;
     using Rules.Framework.Core.ConditionNodes;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class DecimalConditionNodeTests
