+++ tests/Rules.Framework.Tests/Core/ContentContainerTests.cs
@@ -3,7 +3,7 @@ namespace Rules.Framework.Tests.Core
     using System;
     using FluentAssertions;
     using Rules.Framework.Core;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class ContentContainerTests
