+++ tests/Rules.Framework.Tests/Core/RuleTests.cs
@@ -4,7 +4,7 @@ namespace Rules.Framework.Tests.Core
     using FluentAssertions;
     using Moq;
     using Rules.Framework.Core;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class RuleTests
