+++ tests/Rules.Framework.Tests/Evaluation/ConditionsEvalEngineTests.cs
@@ -8,7 +8,7 @@ namespace Rules.Framework.Tests.Evaluation
     using Rules.Framework.Core.ConditionNodes;
     using Rules.Framework.Evaluation;
     using Rules.Framework.Evaluation.ValueEvaluation;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class ConditionsEvalEngineTests
