+++ tests/Rules.Framework.Tests/Extensions/GenericRuleExtensionsTests.cs
@@ -8,7 +8,7 @@ namespace Rules.Framework.Tests.Extensions
     using Rules.Framework.Core.ConditionNodes;
     using Rules.Framework.Extensions;
     using Rules.Framework.Generics;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class GenericRuleExtensionsTests
