+++ tests/Rules.Framework.Tests/Extensions/GenericSearchArgsExtensionsTests.cs
@@ -5,7 +5,7 @@ namespace Rules.Framework.Tests.Extensions
     using FluentAssertions;
     using Rules.Framework.Extensions;
     using Rules.Framework.Generics;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class GenericSearchArgsExtensionsTests
