+++ tests/Rules.Framework.Tests/Generics/GenericRulesEngineTests.cs
@@ -8,7 +8,7 @@ namespace Rules.Framework.Tests.Generics
     using Rules.Framework.Core;
     using Rules.Framework.Core.ConditionNodes;
     using Rules.Framework.Generics;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Xunit;
 
     public class GenericRulesEngineTests
