+++ tests/Rules.Framework.Tests/RulesEngineTests.cs
@@ -13,7 +13,7 @@ namespace Rules.Framework.Tests
     using Rules.Framework.Core.ConditionNodes;
     using Rules.Framework.Evaluation;
     using Rules.Framework.Source;
-    using Rules.Framework.Tests.TestStubs;
+    using Rules.Framework.Tests.Stubs;
     using Rules.Framework.Validation;
     using Xunit;
