+++ tests/Rules.Framework.Tests/TestStubs/ConditionTypeClass.cs
@@ -1,4 +1,4 @@
-namespace Rules.Framework.Tests.TestStubs
+namespace Rules.Framework.Tests.Stubs
 {
     internal class ConditionTypeClass
     {
