+++ tests/Rules.Framework.Tests/TestStubs/ContentTypeClass.cs
@@ -1,4 +1,4 @@
-namespace Rules.Framework.Tests.TestStubs
+namespace Rules.Framework.Tests.Stubs
 {
     internal class ContentTypeClass
     {
